[all] signedness, narrowing conversion fixes

+++ dune/stuff/function/interface.hh
@@ -50,14 +50,14 @@ public:
     // process input
     assert(arg.size() == dimDomain);
     DomainType arg_fvector;
-    for (int i = 0; i < dimDomain; ++i)
+    for (size_t i = 0; i < dimDomain; ++i)
       arg_fvector[i] = arg(i);
     // evaluate
     RangeType ret_fvector;
     evaluate(arg_fvector, ret_fvector);
     // process output
     assert(ret.size() == dimRange);
-    for (int i = 0; i < dimRange; ++i)
+    for (size_t i = 0; i < dimRange; ++i)
       ret(i) = ret_fvector[i];
   }
 
@@ -75,7 +75,7 @@ public:
     evaluate(arg, ret_fvector);
     // process output
     assert(ret.size() == dimRange);
-    for (int i = 0; i < dimRange; ++i)
+    for (size_t i = 0; i < dimRange; ++i)
       ret(i) = ret_fvector[i];
   }
 
@@ -84,7 +84,7 @@ public:
     // process input
     assert(arg.size() == dimDomain);
     DomainType arg_fvector;
-    for (int i = 0; i < dimDomain; ++i)
+    for (size_t i = 0; i < dimDomain; ++i)
       arg_fvector[i] = arg(i);
     // evaluate
     RangeType ret_fvector;
